Update PaperColor colorscheme
This commit fixes the configuration for PaperColor so that it can be
properly reloaded after loading a different colorscheme. This is done
using autocommands. This allows custom configuration code to run
immediately before and immediately after loading the PaperColor theme.

The autocommand before loading the colorscheme is useful for setting
options that allow the colorscheme to be displayed properly. The
autocommand afterwards is useful for setting custom color overrides for
the colorscheme.