Move Leader config to options
In general, the leader keys are better set in the options file than the
keymap file. This is primarily because they should be set early in the
config, before lazy loads plugins so that the correct mappings are
assigned when plugins are loaded. This does not need to be in the
options file, specifically, but these options can be safely set before
lazy runs.